Move Phone and Watch implementations to common module
