+++ mobile/mobile-core/src/commonTest/kotlin/com/boswelja/watchconnection/TestHelpers.kt
@@ -1,6 +1,6 @@
 package com.boswelja.watchconnection
 
-import com.boswelja.watchconnection.core.Watch
+import com.boswelja.watchconnection.common.Watch
 
 fun createWatchesFor(count: Int, platformIdentifier: String): List<Watch> {
     return (0 until count).map {
