+++ mobile/platform-tizen/src/androidMain/kotlin/com/boswelja/watchconnection/tizen/TizenDiscoveryPlatform.kt
@@ -1,8 +1,8 @@
 package com.boswelja.watchconnection.tizen
 
 import android.content.Context
+import com.boswelja.watchconnection.common.Watch
 import com.boswelja.watchconnection.common.discovery.Status
-import com.boswelja.watchconnection.core.Watch
 import com.boswelja.watchconnection.core.discovery.DiscoveryPlatform
 import com.samsung.android.sdk.accessory.SAAgentV2
 import kotlinx.coroutines.flow.Flow
