+++ wear/src/androidMain/kotlin/com/boswelja/watchconnection/wear/discovery/DiscoveryClient.kt
@@ -2,7 +2,7 @@ package com.boswelja.watchconnection.wear.discovery
 
 import android.content.Context
 import android.net.Uri
-import com.boswelja.watchconnection.wear.Phone
+import com.boswelja.watchconnection.common.Phone
 import com.boswelja.watchconnection.wear.repeating
 import com.google.android.gms.wearable.CapabilityClient
 import com.google.android.gms.wearable.Wearable
