+++ wear/src/androidMain/kotlin/com/boswelja/watchconnection/wear/message/MessageClient.kt
@@ -1,10 +1,10 @@
 package com.boswelja.watchconnection.wear.message
 
 import android.content.Context
+import com.boswelja.watchconnection.common.Phone
 import com.boswelja.watchconnection.common.message.Message
 import com.boswelja.watchconnection.common.message.ReceivedMessage
 import com.boswelja.watchconnection.common.message.serialized.MessageSerializer
-import com.boswelja.watchconnection.wear.Phone
 import com.google.android.gms.wearable.MessageClient
 import com.google.android.gms.wearable.Wearable
 import kotlinx.coroutines.ExperimentalCoroutinesApi
