--- wear/src/commonMain/kotlin/com/boswelja/watchconnection/wear/Phone.kt
@@ -1,8 +1,0 @@
-package com.boswelja.watchconnection.wear
-
-import com.boswelja.watchconnection.common.Device
-
-public open class Phone(
-    uid: String,
-    name: String,
-) : Device(uid, name)
